[#38] Remove some unneeded code
