--- Survey/Sources/Domain/Entities/AnswerSubmission.swift
@@ -1,15 +1,0 @@
-//
-//  AnswerSubmission.swift
-//  Survey
-//
-//  Created by Khanh on 14/11/2022.
-//  Copyright © 2022 Nimble. All rights reserved.
-//
-
-import Foundation
-
-struct AnswerSubmission {
-
-    var id: String
-    var answer: String?
-}
