+++ Survey/Sources/Domain/Entities/DisplayType.swift
@@ -1,0 +1,23 @@
+//
+//  DisplayType.swift
+//  Survey
+//
+//  Created by Khanh on 18/11/2022.
+//  Copyright © 2022 Nimble. All rights reserved.
+//
+
+import Foundation
+
+enum DisplayType: String, Codable {
+
+    case star
+    case heart
+    case smiley
+    case choice
+    case dropdown
+    case nps
+    case textarea
+    case textfield
+    case intro
+    case outro
+}
