+++ Survey/Sources/Domain/Entities/PickType.swift
@@ -1,0 +1,14 @@
+//
+//  PickType.swift
+//  Survey
+//
+//  Created by Khanh on 18/11/2022.
+//  Copyright © 2022 Nimble. All rights reserved.
+//
+
+import Foundation
+
+enum PickType: String, Codable {
+
+    case one, any, none
+}
