--- Survey/Sources/Domain/Entities/QuestionSubmission.swift
@@ -1,15 +1,0 @@
-//
-//  QuestionSubmission.swift
-//  Survey
-//
-//  Created by Khanh on 14/11/2022.
-//  Copyright © 2022 Nimble. All rights reserved.
-//
-
-import Foundation
-
-struct QuestionSubmission {
-
-    var id: String
-    var answers: [AnswerSubmission]
-}
